+++ gamevault/UserControls/GeneralControls/FileSelectorUserControl.xaml.cs
@@ -101,7 +101,7 @@ namespace gamevault.UserControls
                                     bitmap.Save(pasteStream, ImageFormat.Png);
                                 }
                                 pasteStream.Seek(0, SeekOrigin.Begin);
-                                string resp = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/v1/images", pasteStream, "x.png", null);
+                                string resp = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/images", pasteStream, "x.png", null);
                                 ImageId = JsonSerializer.Deserialize<Image>(resp).ID;
                                 MainWindowViewModel.Instance.AppBarText = "Sucessfully uploaded image";
                             }
@@ -109,7 +109,7 @@ namespace gamevault.UserControls
                         else
                         {
                             ArrayList filesArray = (ArrayList)Clipboard.GetFileDropList().SyncRoot;
-                            string response = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/v1/images", File.OpenRead(filesArray[0].ToString()), System.IO.Path.GetFileName(filesArray[0].ToString()), null);
+                            string response = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/images", File.OpenRead(filesArray[0].ToString()), System.IO.Path.GetFileName(filesArray[0].ToString()), null);
                             ImageId = JsonSerializer.Deserialize<Image>(response).ID;
                             MainWindowViewModel.Instance.AppBarText = "Sucessfully uploaded image";
                         }
@@ -133,7 +133,7 @@ namespace gamevault.UserControls
                     if (result == System.Windows.Forms.DialogResult.OK && File.Exists(dialog.FileName))
                     {
                         IsUploading = true;
-                        string response = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/v1/images", File.OpenRead(dialog.FileName), System.IO.Path.GetFileName(dialog.FileName), null);
+                        string response = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/images", File.OpenRead(dialog.FileName), System.IO.Path.GetFileName(dialog.FileName), null);
                         ImageId = JsonSerializer.Deserialize<Image>(response).ID;
                         MainWindowViewModel.Instance.AppBarText = "Sucessfully uploaded image";
                     }
@@ -155,7 +155,7 @@ namespace gamevault.UserControls
                     string[] files = (string[])e.Data.GetData(DataFormats.FileDrop);
                     var file = files[0];
                     IsUploading = true;
-                    string response = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/v1/images", File.OpenRead(file), System.IO.Path.GetFileName(file), null);
+                    string response = await WebHelper.UploadFileAsync($"{SettingsViewModel.Instance.ServerUrl}/api/images", File.OpenRead(file), System.IO.Path.GetFileName(file), null);
                     ImageId = JsonSerializer.Deserialize<Image>(response).ID;
                     MainWindowViewModel.Instance.AppBarText = "Sucessfully uploaded image";
                 }